+++ test/source.test.ts
@@ -1,6 +1,6 @@
 import { configure } from "mobx";
 import { types, getSnapshot } from "mobx-state-tree";
-import { Source, Form, converters, Field } from "../src";
+import { Source, Form, converters, Field, RepeatingForm } from "../src";
 import { resolveReactions } from "./util";
 
 // "always" leads to trouble during initialization.
@@ -353,6 +353,68 @@ describe("source accessor in fields", () => {
 
     disposeA();
     disposeB();
+  });
+
+  test("form with autoLoad disposes on delete", async () => {
+    const n = types.model("N", {
+      foo: types.array(M),
+      containerA: ContainerA,
+      containerB: ContainerB
+    });
+
+    const r = n.create({
+      foo: [
+        {
+          a: undefined,
+          b: undefined
+        }
+      ],
+      containerA: { entryMap: {} },
+      containerB: { entryMap: {} }
+    });
+
+    const containerA = r.containerA;
+    const containerB = r.containerB;
+
+    const sourceA = new Source({ entryMap: containerA.entryMap, load: loadA });
+    const sourceB = new Source({ entryMap: containerB.entryMap, load: loadB });
+
+    let dependentQueryCounter = 0;
+
+    const form = new Form(n, {
+      foo: new RepeatingForm({
+        a: new Field(converters.maybe(converters.model(ItemA)), {
+          references: {
+            source: sourceA
+          }
+        }),
+        b: new Field(converters.maybe(converters.model(ItemB)), {
+          references: {
+            source: sourceB,
+            // this source is dependent on state. This information
+            // should be sent whenever a load is issued
+            dependentQuery: accessor => {
+              dependentQueryCounter += 1;
+              return { a: accessor.node.a, c: accessor.node.c };
+            }
+          }
+        })
+      })
+    });
+    const state = form.state(r);
+    const repeating = state.repeatingForm("foo");
+    const firstLine = repeating.index(0);
+
+    const fieldA = firstLine.field("a");
+    const fieldB = firstLine.field("b");
+    fieldA.references.autoLoadReaction();
+    fieldB.references.autoLoadReaction();
+
+    expect(dependentQueryCounter).toEqual(1);
+
+    repeating.remove(r.foo[0]);
+
+    expect(dependentQueryCounter).toEqual(1);
   });
 
   test("no references", async () => {